test: consolidate and fix session runtime tests
